move components to style guide and remove unused code
